requirements.txt
Django==5.2.6
djangorestframework==3.15.2
mysqlclient==2.2.4
requests==2.32.3
Pillow==10.4.0
python-decouple==3.8
tweepy==4.14.0